backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
opencv-python
pillow
numpy
pytest
python-multipart
python-dotenv
requests
